backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
werkzeug==2.0.3
jinja2==3.0.1
itsdangerous==2.0.1
click==8.0.1
gunicorn==20.1.0
python-dotenv==0.19.2